Rename permute_indices.h → permute_indices_wrapper.h (closes #183)
The header `permute_indices.h` defined a class `permute_indices_wrapper`
— the only outright header-vs-class name mismatch in
`include/numsim_cas/tensor/wrappers/`. Issue #183 surveyed the folder's
naming conventions and option 1 (match header to class) was the
least-invasive fix.

This commit renames just that one file via `git mv` and updates the
four headers that include it: skew_classification.h,
tensor_data_inner_product.h, tensor_evaluator.h,
tensor_definitions.h.

Other naming inconsistencies in the folder (tensor_inv.h /
tensor_pow.h not carrying _wrapper despite being wrappers) are
preserved as-is — those headers' classes also lack the suffix, so
header and class names already match. No change to public API.

1157/1157 tests pass.

Closes #183.
