+++ include/numsim_cas/tensor/tensor_definitions.h
@@ -13,7 +13,7 @@
 #include <numsim_cas/tensor/tensor_zero.h>
 #include <numsim_cas/tensor/wrappers/inner_product_wrapper.h>
 #include <numsim_cas/tensor/wrappers/outer_product_wrapper.h>
-#include <numsim_cas/tensor/wrappers/permute_indices.h>
+#include <numsim_cas/tensor/wrappers/permute_indices_wrapper.h>
 #include <numsim_cas/tensor/wrappers/simple_outer_product.h>
 #include <numsim_cas/tensor/wrappers/tensor_inv.h>
 #include <numsim_cas/tensor/wrappers/tensor_pow.h>
